requirements.txt
streamlit==1.32.0
streamlit-reveal-slides==0.2.0
python-dotenv==1.0.0
